next-authのSessionおよびJWTインターフェースにtokenプロパティを追加するモジュールを新規作成
