+++ my-next-app/app/auth/authModule.ts
@@ -1,0 +1,14 @@
+declare module "next-auth" {
+  interface Session {
+    token?: string;
+  }
+}
+
+declare module "next-auth/jwt" {
+  interface JWT {
+    access_token?: string;
+  }
+}
+
+export type { Session } from "next-auth";
+export type { JWT } from "next-auth/jwt";
